+++ src/atoms/exp_cone/log.jl
@@ -15,7 +15,7 @@ type LogAtom <: AbstractExpr
   head::Symbol
   id_hash::Uint64
   children::(AbstractExpr,)
-  size::(Int64, Int64)
+  size::(Int, Int)
 
   function LogAtom(x::AbstractExpr)
     children = (x,)
