+++ src/atoms/maximum.jl
@@ -12,7 +12,7 @@ type MaximumAtom <: AbstractExpr
   head::Symbol
   id_hash::Uint64
   children::(AbstractExpr,)
-  size::(Int64, Int64)
+  size::(Int, Int)
 
   function MaximumAtom(x::AbstractExpr)
     children = (x,)
